hlc_project/requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
ruptures>=1.1.7
Pillow>=10.0.0
networkx>=3.1
